backend/chat-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
qdrant-client==1.7.0
httpx==0.25.2